+++ src/Rector/Set/DoctrineEncryptionSetList.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kyzegs\DoctrineEncryptionBundle\Rector\Set;
+
+final class DoctrineEncryptionSetList
+{
+    public const MIGRATE_FROM_SPEC_SHAPER_ENCRYPT_BUNDLE = __DIR__.'/../../../rector/sets/mogilvie-encrypt-bundle-to-doctrine-encryption.php';
+}